Refactor core models
