--- invoice-manager-ui/src/app/models/catalogos/giro.ts
@@ -1,6 +1,0 @@
-export class Giro{
-
-    public clave:number;
-    public nombre:string;
-    public fechaActualizacion:Date;
-}
