--- invoice-manager-ui/src/app/models/factura/status-factura.ts
@@ -1,10 +1,0 @@
-export class StatusFactura{
-
-    public id : number;
-    public statusEvento : string ;
-    public statusPago : string;
-    public statusDevolucion : string;
-    public fechaCreacion : Date
-    public fechaActualizacion : Date;
-
-}
