--- invoice-manager-ui/src/app/models/status-factura.ts
@@ -1,12 +1,0 @@
-export class StatusFactura{
-
-    private id : number;
-	private statusEvento : string;
-	private statusPago : string;
-	private statusDevolucion : string;
-	private fechaActualizacion: Date;
-
-	constructor(id?:number){
-		this.id = id || 0;
-	}
-}
