--- invoice-manager-ui/src/app/models/user.ts
@@ -1,19 +1,0 @@
-import { NbMenuItem } from '@nebular/theme';
-import { Role } from './role';
-
-export class  User {
-    public email: string;
-    public activo: boolean;
-    public name: string;
-    public urlPicture: string;
-    public roles: Role[];
-    public menu: NbMenuItem[];
-    public id: number;
-    public alias: string;
-
-
-    constructor() {
-        this.activo = false;
-        this.menu = [];
-    }
-}
