+++ invoice-manager-ui/src/app/pages/contabilidad/pre-cfdi/pre-cfdi.component.ts
@@ -7,7 +7,7 @@ import { Empresa } from '../../../models/empresa';
 import { ClaveProductoServicio } from '../../../models/catalogos/producto-servicio';
 import { ClaveUnidad } from '../../../models/catalogos/clave-unidad';
 import { UsoCfdi } from '../../../models/catalogos/uso-cfdi';
-import { Factura } from '../../../models/factura/factura';
+import { Factura } from '../../../models/factura';
 import { InvoicesData } from '../../../@core/data/invoices-data';
 import { PagoBase } from '../../../models/pago-base';
 import { ActivatedRoute, Router } from '@angular/router';
@@ -21,7 +21,7 @@ import { PaymentsData } from '../../../@core/data/payments-data';
 import { ClientsData } from '../../../@core/data/clients-data';
 import { GenericPage } from '../../../models/generic-page';
 import { Client } from '../../../models/client';
-import { User } from '../../../models/user';
+import { User } from '../../../@core/models/user';
 import { CfdiData } from '../../../@core/data/cfdi-data';
 import { Concepto } from '../../../@core/models/cfdi/concepto';
 import { Pago } from '../../../@core/models/cfdi/pago';
